Implement sending XposedService
