+++ manager/src/main/java/org/lsposed/lspatch/manager/XposedService.kt
@@ -1,0 +1,8 @@
+package org.lsposed.lspatch.manager
+
+import io.github.xposed.xposedservice.IXposedService
+
+class XposedService : IXposedService.Stub() {
+
+    override fun getVersion() = 100
+}
